+++ electron/main.js
@@ -1,4 +1,6 @@
-const {app, BrowserWindow} = require('electron')
+const {app} = require('electron')
+
+const AspectRatioBrowserWindow = require('electron-aspect-ratio-browser-window')
 
 const os = require('os'),
   package = require('../package.json'),
@@ -16,16 +18,20 @@ if (os.platform() == 'win32') {
 }
 
 function createWindow() {
-  mainWindow = new BrowserWindow({
+  mainWindow = new AspectRatioBrowserWindow({
     frame: false,
     fullscreen: true,
     icon: path.join(__dirname, '../public/favicon.png'),
     title: package.name,
+    height: 64,
+    width: 64,
     webPreferences: {
       devTools: true,
       preload: path.join(__dirname, 'preload.js'),
     }
   })
+
+  mainWindow.setAspectRatio(1)
 
   mainWindow.webContents.session.setPermissionRequestHandler((webContents, permission, callback) => {
     switch (permission) {