+++ docs/CHANGELOG.md
@@ -3,5 +3,6 @@
 ## v1.1.0
 - Added mouse controls
 - Added user documentation
+- Enforced desktop aspect ratio while in windowed mode
 - Standardized animations across refresh rates
 - Upgraded to latest version of syngen